add final cleanup
