+++ src/components/auth/components/Recaptcha/Recaptcha.native.js
@@ -1,7 +1,11 @@
 import React, { forwardRef, useCallback, useImperativeHandle, useRef } from 'react'
 import ReCAPTCHA from 'react-native-recaptcha-that-works'
 import { noop } from 'lodash'
-const Recaptcha = forwardRef(({ siteKey, baseUrl, onError, onVerify, children, ...props }, ref) => {
+import Config from '../../../../config/config'
+
+const { recaptchaSiteKey, baseUrl } = Config
+
+const Recaptcha = forwardRef(({ onError, onVerify, children, ...props }, ref) => {
   const captchaRef = useRef()
   const onExpired = useCallback(() => captchaRef.current.close(), [])
 
@@ -18,7 +22,7 @@ const Recaptcha = forwardRef(({ siteKey, baseUrl, onError, onVerify, children, .
       <ReCAPTCHA
         {...props}
         ref={captchaRef}
-        siteKey={siteKey}
+        siteKey={recaptchaSiteKey}
         baseUrl={baseUrl}
         size="normal"
         onExpire={onExpired}
