+++ src/components/permissions/api/PermissionsAPI.web.js
@@ -1,11 +1,9 @@
 // @flow
 import { invokeMap, isFunction } from 'lodash'
-import Config from '../../../config/config'
 
 import logger from '../../../lib/logger/js-logger'
 import { type Permission, Permissions, type PermissionStatus, PermissionStatuses } from '../types'
 
-const { enableWebNotifications } = Config
 class PermissionsAPI {
   // permissions enum to platform permissions map
   platformPermissions = {
@@ -14,7 +12,7 @@ class PermissionsAPI {
   }
 
   disabledPermissions = {
-    [Permissions.Notifications]: !enableWebNotifications,
+    [Permissions.Notifications]: true,
   }
 
   constructor(api, clipboardApi, mediaApi, log) {
